+++ src/services/users.services.ts
@@ -8,13 +8,24 @@ import { envConfig } from '~/constants/config'
 import RefreshToken from '~/models/schemas/RefreshToken.schema'
 import { ObjectId } from 'mongodb'
 import { USERS_MESSAGES } from '~/constants/messages'
-import { omit } from 'lodash'
+import { isNumber, omit } from 'lodash'
 import { ErrorWithStatus } from '~/models/Errors'
 import HTTP_STATUS from '~/constants/httpStatus'
 import axios from 'axios'
 import { sendVerifyRegisterEmail, sendForgotPasswordEmail } from '~/utils/email'
 import Company from '~/models/schemas/Company.schema'
 import { PositionType } from '~/models/requests/Company.request'
+import { io } from '..'
+import { activeConnections } from '~/app/socket'
+
+export interface QueryUserEmployerFilter {
+  content?: string
+  verify?: string
+  status?: string
+  limit?: string
+  page?: string
+  type?: 'employer' | 'admin' | 'candidate' | 'all'
+}
 
 class UsersService {
   private signAccessToken({ user_id, verify, role }: { user_id: string; verify: UserVerifyStatus; role: UserRole }) {
@@ -117,7 +128,8 @@ class UsersService {
       await databaseServices.company.insertOne(
         new Company({
           company_name: payload.company_name as string,
-          working_locations: [],
+          fields: payload.fields || [],
+          working_locations: payload.working_locations || [],
           users: [
             {
               user_id,
@@ -131,7 +143,7 @@ class UsersService {
     const [access_token, refresh_token] = await this.signTokenKeyPair({
       user_id: user_id.toString(),
       verify: UserVerifyStatus.Unverified,
-      role: payload.role
+      role: payload.role || UserRole.Candidate
     })
     const { iat, exp } = await this.decodeRefreshToken(refresh_token)
     await databaseServices.refreshTokens.insertOne(new RefreshToken({ user_id, token: refresh_token, iat, exp }))
@@ -143,7 +155,7 @@ class UsersService {
     // 4. Server verify email_verify_token
     // 5. Client receive access_token and refresh_token
     // console.log('gửi token verification email:', email_verify_token)
-    // await sendVerifyRegisterEmail(payload.email, email_verify_token)
+    await sendVerifyRegisterEmail(payload.email, email_verify_token)
 
     return {
       access_token,
@@ -276,18 +288,26 @@ class UsersService {
     user_id,
     verify,
     refresh_token,
-    exp,
-    role
+    exp
   }: {
     user_id: string
     verify: UserVerifyStatus
     refresh_token: string
     exp: number
-    role: UserRole
   }) {
+    const user = await databaseServices.users.findOne({
+      _id: new ObjectId(user_id)
+    })
+
+    if (!user)
+      throw new ErrorWithStatus({
+        message: 'User not found',
+        status: 404
+      })
+
     const [new_access_token, new_refresh_token] = await Promise.all([
-      this.signAccessToken({ user_id, verify, role }),
-      this.signRefreshToken({ user_id, verify, exp })
+      this.signAccessToken({ user_id, verify: user.verify, role: user.role }),
+      this.signRefreshToken({ user_id, verify: user.verify, exp })
     ])
 
     await databaseServices.refreshTokens.deleteOne({ token: refresh_token })
@@ -331,7 +351,7 @@ class UsersService {
         user_id: new ObjectId(user_id)
       })
     )
-
+    io.to(user_id).emit('verify', {})
     return { access_token, refresh_token }
   }
 
@@ -498,6 +518,133 @@ class UsersService {
       user_id: new ObjectId(userId)
     })
     return result
+  }
+
+  async getUsersFromAdmin(filter: QueryUserEmployerFilter) {
+    const limit = Number(filter.limit) || 10
+    const page = Number(filter.page) || 1
+
+    const opts = this.queryUserEmployerConvertToMatch(filter) || {}
+
+    console.log(opts)
+
+    const [employers, total] = await Promise.all([
+      databaseServices.users
+        .aggregate([
+          {
+            $match: {
+              ...opts
+            }
+          },
+          {
+            $project: {
+              password: 0,
+              email_verify_token: 0,
+              forgot_password_token: 0
+            }
+          },
+          {
+            $skip: limit * (page - 1)
+          },
+          {
+            $limit: limit
+          }
+        ])
+        .toArray(),
+      databaseServices.users
+        .aggregate([
+          {
+            $match: {
+              role: UserRole.Employer,
+              ...opts
+            }
+          },
+
+          {
+            $count: 'total'
+          }
+        ])
+        .toArray()
+    ])
+
+    return {
+      employers,
+      total: total[0]?.total || 0,
+      limit,
+      page
+    }
+  }
+
+  queryUserEmployerConvertToMatch(query: QueryUserEmployerFilter) {
+    const options: {
+      [key: string]: any
+    } = {}
+    if (query.content) {
+      options['$text'] = {
+        $search: query.content
+      }
+    }
+
+    if (query.status && isNumber(Number(query.status))) {
+      options['status'] = Number(query.status)
+    }
+
+    if (query.verify && isNumber(Number(query.verify))) {
+      options['verify'] = Number(query.verify)
+    }
+
+    if (query.type) {
+      if (query.type === 'admin') options['role'] = UserRole.Administrators
+      if (query.type === 'candidate') options['role'] = UserRole.Candidate
+      if (query.type === 'employer') options['role'] = UserRole.Employer
+    }
+    return options
+  }
+
+  async lockOrUnlockUser(userId: string) {
+    const user = await databaseServices.users.findOne({
+      _id: new ObjectId(userId),
+      role: {
+        $in: [UserRole.Candidate, UserRole.Employer]
+      }
+    })
+
+    if (!user || user.verify === UserVerifyStatus.Unverified) {
+      return false
+    }
+
+    if (user.status === 0) {
+      await databaseServices.users.updateOne(
+        {
+          _id: user._id
+        },
+        {
+          $set: {
+            status: 1,
+            verify: UserVerifyStatus.Banned
+          }
+        }
+      )
+
+      await databaseServices.refreshTokens.deleteMany({
+        user_id: user._id
+      })
+      io.to(user._id.toString()).emit('lock-user', {})
+    } else {
+      await databaseServices.users.updateOne(
+        {
+          _id: user._id
+        },
+        {
+          $set: {
+            status: 0,
+            verify: UserVerifyStatus.Verified
+          }
+        }
+      )
+    }
+
+    return true
   }
 }
